new upgrade test
