--- src/test/java/BasicTest.java
@@ -1,25 +1,0 @@
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.testng.annotations.AfterTest;
-import org.testng.annotations.BeforeTest;
-
-import java.util.concurrent.TimeUnit;
-
-public class BasicTest {
-    WebDriver driver;
-
-    @BeforeTest
-    public void setUp() {
-        System.setProperty("webdriver.chrome.driver", "src/test/resources/chromedriver.exe");
-        ChromeOptions options = new ChromeOptions();
-        options.addArguments("--start-maximized");
-        driver = new ChromeDriver(options);
-        driver.manage().timeouts().implicitlyWait(60, TimeUnit.SECONDS);
-    }
-
-    @AfterTest(alwaysRun = true)
-    public void tearDown() {
-        driver.quit();
-    }
-}
