+++ src/test/java/pages/BasePage.java
@@ -1,0 +1,22 @@
+package pages;
+
+import io.github.bonigarcia.wdm.WebDriverManager;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
+
+import java.util.concurrent.TimeUnit;
+
+public class BasePage {
+    WebDriver driver;
+
+    public BasePage(WebDriver driver){
+     this.driver= driver;
+
+    }
+
+}
+
+
