+++ src/test/java/pages/LoginPage.java
@@ -1,0 +1,30 @@
+package pages;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+
+public class LoginPage extends BasePage  {
+
+
+    public  static final By USERNAME_INPUT = By.xpath("//input[@name='user-name']");
+    public static  final By PASSWORD_INPUT = By.id("password");
+    public static final  By LOGIN_BUTTON = By.cssSelector("[name='login-button']");
+    public static final By ERROR_MESSAGE = By.cssSelector(".error_message_container");
+
+
+    public LoginPage(WebDriver driver){
+                super(driver);
+    }
+   public void  open(){
+     driver.get("https://www.saucedemo.com/");
+    }
+    public void login(String userName, String password){
+        driver.findElement(USERNAME_INPUT).sendKeys(userName);
+        driver.findElement(PASSWORD_INPUT).sendKeys(password);
+        driver.findElement(LOGIN_BUTTON).click();
+
+    }
+    public  String getErrorMessage(){
+       return driver.findElement(ERROR_MESSAGE).getText();
+    }
+}
