+++ src/test/java/tests/BaseTest.java
@@ -1,0 +1,41 @@
+package tests;
+
+import io.github.bonigarcia.wdm.WebDriverManager;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.AfterTest;
+import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.BeforeTest;
+import pages.CheckoutYourInformation;
+import pages.InventoryPage;
+import pages.LoginPage;
+
+import java.util.concurrent.TimeUnit;
+
+public class BaseTest {
+    WebDriver driver;
+    LoginPage loginPage;
+    InventoryPage inventoryPage;
+    CheckoutYourInformation checkoutYourInformation;
+
+
+    @BeforeMethod
+    public void setup() {
+        WebDriverManager.chromedriver().setup();
+        ChromeOptions options = new ChromeOptions();
+        options.addArguments("--start-maximized");
+        driver = new ChromeDriver(options);
+        driver.manage().timeouts().implicitlyWait(20, TimeUnit.SECONDS);
+
+        loginPage = new LoginPage(driver);
+        inventoryPage = new InventoryPage(driver);
+        checkoutYourInformation = new CheckoutYourInformation(driver);
+    }
+
+    @AfterMethod(alwaysRun = true)
+    public void tearDown() {
+        //driver.quit();
+    }
+}
